requirements.txt
gTTS==2.1.1
PyAudio==0.2.11
SpeechRecognition==3.8.1
playsound==1.2.2
python-decouple==3.4
